+++ src/components/composite/ButtonGroup/index.ts
@@ -22,5 +22,6 @@ export {
   ButtonGroupClassNamesProvider,
   useOptionalButtonGroupLayout,
   useOptionalButtonGroupSegment,
+  useInJoinedButtonGroup,
   useButtonGroupClassNames,
 } from "./buttonGroupContext";